refactor: extract shared MoreMenu component
SearchBar.vue and MainMenu.vue had identical more-menu implementations
(trigger, language switcher, about button, keyboard nav, click-outside).
Extract into a shared MoreMenu.vue component with:

- teleport prop: when true, panel renders at <body> with fixed positioning
  (needed for SearchBar's overflow:hidden context)
- @about emit: parent owns the AboutModal lifecycle
- All keyboard navigation, focus management, click-outside handling
- Language list with auto-translation and project store reload on switch
- Merged CSS from both original components
